Fix test_trunk_elevation_zero_norm_vector_raises()
which was not supposed to raise with neuron NRN but since the soma center
was wrong, it was raising.
Soma center was wrong because it is a contour and not a 3 point soma, so
the center is the average of the soma points. And it is not the first point.